Modularization of the connection lines component
